services/document-processor/requirements.txt
redis==5.0.1
fastapi==0.109.0
uvicorn[standard]==0.27.0
boto3==1.34.34
python-dotenv==1.0.0
structlog==24.1.0
